Add files via upload
Attention based LSTM 

Classification model: 

RNN can deal with sequential data and can be trained to hold the information of the input from previous layers. It saves the output of a layer and feeds it back to the input to predict the output of the layer. But it can only hold the data for a short term. LSTM can handle the information in memory for a long period of time as compared to RNN. 

LSTM has multiple layers that hold more memory while passing through each layer, only relevant info is kept and irrelevant info will be discarded. LSTM efficiently improves performance by memorizing the relevant information that is important and finding the pattern.

If we compare LSTM to other classification algorithms, they train the model by considering multiple words as separate inputs that don’t have actual meaning as a sentence. So during the prediction, it gives output based on statistics not on the meaning. By using appropriate layers of embedding and encoding in LSTM. The model can find out the actual meaning of the input and predict the most accurate output class. But this approach compresses all the important information of a sentence into a fixed length vector, this may make the neural network difficult to efficiently cope up with long sentences. To avoid this, attention based architecture is used which attends every hidden state from each encoder node at every step and then makes predictions after deciding which one is more informative. 

Bi-LSTM classification model can be used to process the data both in forward and backward direction.